requirements.txt
anyio==3.6.1
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
fastapi==0.80.0
gunicorn==20.1.0
h11==0.13.0
httptools==0.4.0
idna==3.3
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pydantic==1.9.2
pyparsing==3.0.9
pytest==7.1.2
python-dotenv==0.20.0
PyYAML==6.0
requests==2.28.1
sniffio==1.2.0
starlette==0.19.1
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
uvloop==0.16.0
watchfiles==0.16.1
websockets==10.3
